refactor: 规范化 Pathing 命名与脚本结构
按规范重命名 Pathing 路线并同步 info.name；
移除 Pathing 目录 README，将必要说明合并至项目 README；
拆分 main.js 模块，保留原路线坐标与执行逻辑。
