+++ repo/js/IceObjectNpcLike/Assets/Pathing/00-恢复热能-至冬-七天神像.json
@@ -7,12 +7,12 @@
       }
     ],
     "bgi_version": "0.64.0",
-    "description": "回复热能值",
+    "description": "恢复热能值",
     "enable_monster_loot_split": false,
     "last_modified_time": 1788353423190,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "restore",
+    "name": "00-恢复热能-至冬-七天神像",
     "tags": [],
     "type": "collect",
     "version": "1.0"
