+++ repo/js/IceObjectNpcLike/Assets/Pathing/01-冰造物点赞-至冬-北国银行门口-6赞.json
@@ -7,17 +7,17 @@
       }
     ],
     "bgi_version": "0.64.0",
-    "description": "车站餐厅1人",
+    "description": "北国银行门口3人",
     "enable_monster_loot_split": false,
-    "last_modified_time": 1788415619283,
+    "last_modified_time": 1788355062407,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "point07",
+    "name": "01-冰造物点赞-至冬-北国银行门口-6赞",
     "tags": [
-      "点赞=2"
+      "点赞=6"
     ],
     "type": "collect",
-    "version": "1.0"
+    "version": "2.0"
   },
   "positions": [
     {
@@ -26,8 +26,8 @@
       "id": 1,
       "move_mode": "walk",
       "type": "teleport",
-      "x": 7813.7715,
-      "y": 8520.4678
+      "x": 6583.8818,
+      "y": 9576.1563
     },
     {
       "action": "",
@@ -35,35 +35,26 @@
       "id": 2,
       "move_mode": "walk",
       "type": "path",
-      "x": 7828.5762,
-      "y": 8517.2344
+      "x": 6578.7275,
+      "y": 9599.0156
     },
     {
       "action": "",
       "action_params": "",
       "id": 3,
       "move_mode": "walk",
-      "type": "path",
-      "x": 7833.4082,
-      "y": 8511.2188
+      "type": "target",
+      "x": 6561.6884,
+      "y": 9596.9531
     },
     {
       "action": "",
       "action_params": "",
       "id": 4,
       "move_mode": "walk",
-      "type": "target",
-      "x": 7839.937,
-      "y": 8497.9844
-    },
-    {
-      "action": "",
-      "action_params": "",
-      "id": 5,
-      "move_mode": "walk",
       "type": "orientation",
-      "x": 7799.9284,
-      "y": 8457.25
+      "x": 6579.8499,
+      "y": 9479.9063
     }
   ]
 }
