+++ repo/js/IceObjectNpcLike/Assets/Pathing/03-冰造物点赞-至冬-炉灰食堂门口-4赞.json
@@ -7,17 +7,17 @@
       }
     ],
     "bgi_version": "0.64.0",
-    "description": "北国银行门口3人",
+    "description": "炉灰食堂门口2人",
     "enable_monster_loot_split": false,
-    "last_modified_time": 1788355062407,
+    "last_modified_time": 1788491253964,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "point01",
+    "name": "03-冰造物点赞-至冬-炉灰食堂门口-4赞",
     "tags": [
-      "点赞=6"
+      "点赞=4"
     ],
     "type": "collect",
-    "version": "2.0"
+    "version": "3.0"
   },
   "positions": [
     {
@@ -26,8 +26,8 @@
       "id": 1,
       "move_mode": "walk",
       "type": "teleport",
-      "x": 6583.8818,
-      "y": 9576.1563
+      "x": 6864.9617,
+      "y": 9655.6484
     },
     {
       "action": "",
@@ -35,8 +35,8 @@
       "id": 2,
       "move_mode": "walk",
       "type": "path",
-      "x": 6578.7275,
-      "y": 9599.0156
+      "x": 6860.751,
+      "y": 9644.3906
     },
     {
       "action": "",
@@ -44,8 +44,8 @@
       "id": 3,
       "move_mode": "walk",
       "type": "target",
-      "x": 6561.6884,
-      "y": 9596.9531
+      "x": 6863.1557,
+      "y": 9635.625
     },
     {
       "action": "",
@@ -53,8 +53,8 @@
       "id": 4,
       "move_mode": "walk",
       "type": "orientation",
-      "x": 6579.8499,
-      "y": 9479.9063
+      "x": 6861.0985,
+      "y": 9565.1563
     }
   ]
 }
