+++ repo/js/IceObjectNpcLike/Assets/Pathing/04-冰造物点赞-至冬-餐厅门口-2赞.json
@@ -12,7 +12,7 @@
     "last_modified_time": 1788491196629,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "point04",
+    "name": "04-冰造物点赞-至冬-餐厅门口-2赞",
     "tags": [
       "点赞=2"
     ],
