+++ repo/js/IceObjectNpcLike/Assets/Pathing/05-冰造物点赞-至冬-铁匠附近-2赞.json
@@ -7,17 +7,17 @@
       }
     ],
     "bgi_version": "0.64.0",
-    "description": "车站台子1人",
+    "description": "铁匠",
     "enable_monster_loot_split": false,
-    "last_modified_time": 1788415794723,
+    "last_modified_time": 1788353313560,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "point09_continue",
+    "name": "05-冰造物点赞-至冬-铁匠附近-2赞",
     "tags": [
       "点赞=2"
     ],
     "type": "collect",
-    "version": "1.0"
+    "version": "2.0"
   },
   "positions": [
     {
@@ -25,9 +25,9 @@
       "action_params": "",
       "id": 1,
       "move_mode": "walk",
-      "type": "path",
-      "x": 7822.0635,
-      "y": 8455.875
+      "type": "teleport",
+      "x": 6688.624,
+      "y": 9624.9688
     },
     {
       "action": "",
@@ -35,8 +35,8 @@
       "id": 2,
       "move_mode": "walk",
       "type": "path",
-      "x": 7813.1378,
-      "y": 8455.875
+      "x": 6682.7825,
+      "y": 9607.4375
     },
     {
       "action": "",
@@ -44,8 +44,8 @@
       "id": 3,
       "move_mode": "walk",
       "type": "path",
-      "x": 7811.7338,
-      "y": 8469.2696
+      "x": 6681.9992,
+      "y": 9592.1406
     },
     {
       "action": "",
@@ -53,8 +53,8 @@
       "id": 4,
       "move_mode": "walk",
       "type": "path",
-      "x": 7796.6094,
-      "y": 8470.6563
+      "x": 6683.2895,
+      "y": 9585.8672
     },
     {
       "action": "",
@@ -62,8 +62,8 @@
       "id": 5,
       "move_mode": "walk",
       "type": "target",
-      "x": 7796.5229,
-      "y": 8474.4375
+      "x": 6686.9834,
+      "y": 9580.625
     },
     {
       "action": "",
@@ -71,8 +71,8 @@
       "id": 6,
       "move_mode": "walk",
       "type": "orientation",
-      "x": 7789.0467,
-      "y": 8495.3203
+      "x": 6706.4021,
+      "y": 9558.4531
     }
   ]
 }
