+++ repo/js/IceObjectNpcLike/Assets/Pathing/06-冰造物点赞-至冬-曙光车站玩具附近-4赞.json
@@ -6,18 +6,18 @@
         "name": "一只低熵体"
       }
     ],
-    "bgi_version": "0.45.0",
-    "description": "车站火炬边上3人",
+    "bgi_version": "0.64.0",
+    "description": "曙光车站玩具2人",
     "enable_monster_loot_split": false,
-    "last_modified_time": 1788489561244,
+    "last_modified_time": 1788354846394,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "point08_continue",
+    "name": "06-冰造物点赞-至冬-曙光车站玩具附近-4赞",
     "tags": [
-      "点赞=6"
+      "点赞=4"
     ],
     "type": "collect",
-    "version": "1.0"
+    "version": "3.0"
   },
   "positions": [
     {
@@ -25,9 +25,9 @@
       "action_params": "",
       "id": 1,
       "move_mode": "walk",
-      "type": "path",
-      "x": 7839.937,
-      "y": 8497.9844
+      "type": "teleport",
+      "x": 7813.7715,
+      "y": 8520.4678
     },
     {
       "action": "",
@@ -35,26 +35,35 @@
       "id": 2,
       "move_mode": "walk",
       "type": "path",
-      "x": 7823.9371,
-      "y": 8461.6328
+      "x": 7808.5151,
+      "y": 8513.9688
     },
     {
       "action": "",
       "action_params": "",
       "id": 3,
-      "move_mode": "jump",
-      "type": "target",
-      "x": 7829.6087,
-      "y": 8455.1016
+      "move_mode": "walk",
+      "type": "path",
+      "x": 7793.3958,
+      "y": 8517.4063
     },
     {
       "action": "",
       "action_params": "",
       "id": 4,
       "move_mode": "walk",
+      "type": "target",
+      "x": 7792.365,
+      "y": 8511.5625
+    },
+    {
+      "action": "",
+      "action_params": "",
+      "id": 5,
+      "move_mode": "walk",
       "type": "orientation",
-      "x": 7875.4996,
-      "y": 8473.4063
+      "x": 7727.0131,
+      "y": 8472.375
     }
   ]
 }
