+++ repo/js/IceObjectNpcLike/Assets/Pathing/07-冰造物点赞-至冬-车站餐厅-2赞.json
@@ -7,17 +7,17 @@
       }
     ],
     "bgi_version": "0.64.0",
-    "description": "铁匠",
+    "description": "车站餐厅1人",
     "enable_monster_loot_split": false,
-    "last_modified_time": 1788353313560,
+    "last_modified_time": 1788415619283,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "point05",
+    "name": "07-冰造物点赞-至冬-车站餐厅-2赞",
     "tags": [
       "点赞=2"
     ],
     "type": "collect",
-    "version": "2.0"
+    "version": "1.0"
   },
   "positions": [
     {
@@ -26,8 +26,8 @@
       "id": 1,
       "move_mode": "walk",
       "type": "teleport",
-      "x": 6688.624,
-      "y": 9624.9688
+      "x": 7813.7715,
+      "y": 8520.4678
     },
     {
       "action": "",
@@ -35,8 +35,8 @@
       "id": 2,
       "move_mode": "walk",
       "type": "path",
-      "x": 6682.7825,
-      "y": 9607.4375
+      "x": 7828.5762,
+      "y": 8517.2344
     },
     {
       "action": "",
@@ -44,35 +44,26 @@
       "id": 3,
       "move_mode": "walk",
       "type": "path",
-      "x": 6681.9992,
-      "y": 9592.1406
+      "x": 7833.4082,
+      "y": 8511.2188
     },
     {
       "action": "",
       "action_params": "",
       "id": 4,
       "move_mode": "walk",
-      "type": "path",
-      "x": 6683.2895,
-      "y": 9585.8672
+      "type": "target",
+      "x": 7839.937,
+      "y": 8497.9844
     },
     {
       "action": "",
       "action_params": "",
       "id": 5,
       "move_mode": "walk",
-      "type": "target",
-      "x": 6686.9834,
-      "y": 9580.625
-    },
-    {
-      "action": "",
-      "action_params": "",
-      "id": 6,
-      "move_mode": "walk",
       "type": "orientation",
-      "x": 6706.4021,
-      "y": 9558.4531
+      "x": 7799.9284,
+      "y": 8457.25
     }
   ]
 }
