+++ repo/js/IceObjectNpcLike/Assets/Pathing/08-冰造物点赞-至冬-车站火炬边-6赞.json
@@ -6,15 +6,15 @@
         "name": "一只低熵体"
       }
     ],
-    "bgi_version": "0.64.0",
-    "description": "印记商店门口4人",
+    "bgi_version": "0.45.0",
+    "description": "车站火炬边上3人",
     "enable_monster_loot_split": false,
-    "last_modified_time": 1788414454217,
+    "last_modified_time": 1788489561244,
     "map_match_method": "SIFT",
     "map_name": "Teyvat",
-    "name": "point02",
+    "name": "08-冰造物点赞-至冬-车站火炬边-6赞",
     "tags": [
-      "点赞=8"
+      "点赞=6"
     ],
     "type": "collect",
     "version": "1.0"
@@ -25,9 +25,9 @@
       "action_params": "",
       "id": 1,
       "move_mode": "walk",
-      "type": "teleport",
-      "x": 6583.8818,
-      "y": 9576.1563
+      "type": "path",
+      "x": 7839.937,
+      "y": 8497.9844
     },
     {
       "action": "",
@@ -35,44 +35,26 @@
       "id": 2,
       "move_mode": "walk",
       "type": "path",
-      "x": 6581.7969,
-      "y": 9564.8125
+      "x": 7823.9371,
+      "y": 8461.6328
     },
     {
       "action": "",
       "action_params": "",
       "id": 3,
-      "move_mode": "walk",
-      "type": "path",
-      "x": 6587.6406,
-      "y": 9543.3281
+      "move_mode": "jump",
+      "type": "target",
+      "x": 7829.6087,
+      "y": 8455.1016
     },
     {
       "action": "",
       "action_params": "",
       "id": 4,
       "move_mode": "walk",
-      "type": "path",
-      "x": 6555.1563,
-      "y": 9535.7656
-    },
-    {
-      "action": "",
-      "action_params": "",
-      "id": 5,
-      "move_mode": "walk",
-      "type": "target",
-      "x": 6550.0007,
-      "y": 9528.5469
-    },
-    {
-      "action": "",
-      "action_params": "",
-      "id": 6,
-      "move_mode": "walk",
       "type": "orientation",
-      "x": 6525.5932,
-      "y": 9475.2656
+      "x": 7875.4996,
+      "y": 8473.4063
     }
   ]
 }
