+++ repo/js/IceObjectNpcLike/README.md
@@ -2,45 +2,41 @@
 
 作者：一只低熵体  
 版本：`1.1.0`  
-当前全部路线预计总点赞：`36`
+全部路线预计总点赞：`36`
 
-本脚本使用 BetterGI 地图追踪前往至冬 NPC 点位，在固定工位完成两次冰造物放置与点赞等待，并在离开前删除第二个造物、退出造物界面。脚本不读取游戏内存，不使用联网、第三方程序、KeyMouseScript、OCR 或点赞图像识别。
+脚本使用 BetterGI 地图追踪前往至冬 NPC 工位，自动完成两次冰造物放置、点赞等待、删除和最终清理。脚本不读取游戏内存，不联网，也不使用第三方程序、KeyMouseScript、OCR 或点赞图像识别。
 
-## 使用前准备
+## 使用条件
 
 - BetterGI `0.64.0` 或更高版本；
 - 简体中文游戏客户端；
 - 16:9 游戏画面，至少 `1920×1080`；
-- BetterGI 截图方式使用 `BitBlt`；
-- 原神小地图保持“锁定方向”；
-- 原神水平与垂直镜头灵敏度保持默认值 `3`；
-- 保持默认亮度，关闭 HDR、显卡滤镜等会明显改变画面的功能；
-- 冰造物快捷键为键盘 `T`；
-- 已解锁路线涉及的区域与传送锚点；
+- BetterGI 截图方式设为 `BitBlt`；
+- 小地图保持“锁定方向”；
+- 原神水平、垂直镜头灵敏度均保持默认值 `3`；
+- 保持默认亮度并关闭 HDR、显卡滤镜等画面效果；
+- 冰造物快捷键为 `T`，相关区域和传送锚点已解锁；
 - 运行期间不要操作键盘和鼠标。
 
 ## 任务前置
 
-车站区域的路线有以下任务前置：
+- **Point 06、07、09**：需要完成三个聚所任务的共同第四幕“雪后曙光”，可用成就“睁开被逸夜合拢的眼睛”确认。参考：[《雪后曙光》教程](https://www.bilibili.com/video/BV172bX6XEPm/)。
+- **Point 08**：除“雪后曙光”外，还需在现实时间第二天完成“新雪来客”。参考：[《新雪来客》教程](https://www.bilibili.com/video/BV1nqbX6fEDS/)。
 
-- **Point 06、Point 07、Point 09**：需要完成三个聚所任务的共同第四幕 **“雪后曙光”**。可以通过成就 **“睁开被逸夜合拢的眼睛”** 验证任务是否已经完成。教程参考：[影月月《雪后曙光》](https://www.bilibili.com/video/BV172bX6XEPm/)。
-- **Point 08**：除“雪后曙光”外，还需要在完成该任务后的**现实时间第二天**完成 **“新雪来客”**。教程参考：[影月月《新雪来客》](https://www.bilibili.com/video/BV1nqbX6fEDS/)。
+Point 06～09 默认关闭，请根据任务进度启用。
 
-Point 06～09 在脚本设置中默认关闭，请根据自己的任务进度手动启用。
+## 使用方法
 
-## 快速使用
+1. 在 BetterGI 中订阅或导入脚本。
+2. 右键脚本，打开“修改 JS 自定义配置”。
+3. 按任务进度启用路线；未完成前置任务时保持 Point 06～09 关闭。
+4. 将脚本加入调度器并运行。首次测试建议使用“按 Point 编号”。
 
-1. 在 BetterGI 中订阅或导入本脚本。
-2. 右键脚本并打开“修改 JS 自定义配置”。
-3. 未完成“雪后曙光”时，保持 Point 06～09 关闭；只完成“雪后曙光”时，可以启用 Point 06、07、09并保持 Point 08 关闭。
-4. 将脚本加入调度器并开始运行。
-5. 第一次测试建议按 Point 编号执行，并观察每个 Point 的地图追踪、镜头与清理日志。
-
-默认启用 Point 01～05，预计共获得 `22` 赞；完成“雪后曙光”并启用 Point 06、07、09 后预计共获得 `30` 赞；再完成“新雪来客”并启用 Point 08 后，全部路线预计共获得 `36` 赞。
+默认启用 Point 01～05，预计获得 `22` 赞；完成“雪后曙光”并启用 Point 06、07、09 后预计获得 `30` 赞；再启用 Point 08 后，全部路线预计获得 `36` 赞。
 
 ## 路线一览
 
-| 路线 | 地点 | NPC 数量 | 两次放置预计点赞 | 默认状态 | 备注 |
+| 路线 | 地点 | NPC | 预计点赞 | 默认 | 备注 |
 | --- | --- | ---: | ---: | --- | --- |
 | Point 01 | 北国银行门口 | 3 | 6 | 开启 | 独立路线 |
 | Point 02 | 印记商店门口 | 4 | 8 | 开启 | 独立路线 |
@@ -52,163 +48,44 @@ Point 06～09 在脚本设置中默认关闭，请根据自己的任务进度手
 | Point 08 | 车站火炬边 | 3 | 6 | 关闭 | 接续 Point 07；需完成“新雪来客” |
 | Point 09 | 车站台子 | 1 | 2 | 关闭 | 接续路线；需完成“雪后曙光” |
 
-Point 07、08、09 始终按一个不可拆分的连续路线组调度，组内不会插入传送或热能恢复：
+### 车站连续路线
 
-```text
-Point 08 开启：Point 07 → Point 08 → Point 09
-Point 08 关闭：Point 07 → Point 09 完整备用路线
-```
+- Point 08 开启时：`Point 07 → Point 08 → Point 09`；
+- Point 08 关闭时：自动改走 `Point 07 → Point 09` 完整备用路线；
+- Point 07 关闭时：依赖当前位置的 Point 08、09 会自动跳过。
 
-`point09_from07.json` 是 Point 08 关闭时使用的 **Point 09 完整备用路线**：它从 Point 07 的结束点出发，直接到达 Point 09 工位。路线结束后，脚本会执行一次 Point 09 的镜头与冰造物动作，并按 Point 09 计算 `2` 赞；此分支不会再运行 `point09_continue.json`，也不会重复执行 Point 09。Point 07 关闭时，依赖其起始位置的 Point 08 和 Point 09 会自动跳过。
+连续路线组不会被传送或热能恢复流程拆开。
 
-## 自动化流程
+## 运行模式
 
-脚本启动时先传送至固定七天神像，等待热能自动恢复；随后按设置执行已启用路线：
+- **按 Point 编号**：默认模式，适合首次测试；
+- **高点赞优先**：执行全部已启用路线，并按路线组预计点赞量排序；
+- **高收益精简（6赞及以上）**：只执行预计点赞达到 `6` 的路线组。
 
-```text
-地图追踪至工位
-→ target 精确站位
-→ orientation 确定水平视角
-→ 将垂直镜头推到上限并向下移动固定量
-→ 第一次放置并等待点赞
-→ 删除第一次造物
-→ 第二次放置并等待点赞
-→ 最终删除第二次造物
-→ 按 T 退出冰造物界面
-```
+全部路线启用时，高收益精简模式会执行 Point 07～09 连续组（10赞）、Point 02（8赞）和 Point 01（6赞），合计预计 `24` 赞。
 
-脚本正常完成 Point 时不会故意留下冰造物。如果运行中被手动停止或发生报错，请人工确认现场是否残留造物，并按 `T` 退出造物界面。
+## 热能与镜头
 
-## 路线选择与运行模式
+脚本启动时会先返回固定七天神像恢复热能，默认每完成 `4` 个 Point 再恢复一次。若下一个连续组的长度超过剩余额度，会在组开始前提前恢复。
 
-每条路线都有独立开关。关闭路线后，它不会执行，也不会计入热能恢复次数或预计点赞总数。
+每个 Point 到位后，路线末尾的 `orientation` 负责确定水平视角；脚本再分 `4` 次将垂直镜头推到上限，随后向下移动 `7500`。如个别点位需要调整，可在自定义配置中填写水平或垂直覆盖，例如 `02=6100,05=5900`。
 
-路线运行模式有三种：
-
-- **按 Point 编号**：默认选项，便于测试和定位问题；
-- **高点赞优先**：全部启用路线都会执行，按路线组预计总点赞量从高到低排序；
-- **高收益精简（6赞及以上）**：只执行预计总点赞量达到 `6` 的路线组，并按总点赞量从高到低排序。
-
-筛选和排序都以“可独立调度的路线组”为单位。Point 07、08、09 不会被拆开分别判断：三点全部启用时，该组预计 `10` 赞，会进入高收益精简模式；Point 08 关闭、改走 `Point 07 → Point 09` 备用分支时，该组实际预计 `4` 赞，因此会被高收益精简模式整体过滤。
-
-全部 Point 启用时，高收益精简模式恰好执行三组：
-
-```text
-Point 07 → Point 08 → Point 09：10赞
-Point 02：8赞
-Point 01：6赞
-```
-
-合计预计 `24` 赞。这里的“高收益”只根据点赞总量筛选，没有记录路线耗时，因此不代表严格的单位时间效率。
-
-## 热能恢复
-
-`restore.json` 传送到固定七天神像后会原地自动恢复热能，不需要移动或按 `F`。脚本默认每完成 `4` 个 Point 恢复一次热能，使用约 40 点并保留约 20 点冗余。
-
-连续路线组不会被恢复流程拆开。如果剩余额度不足以执行完整连续组，脚本会在组开始前提前恢复。
-
-## 镜头标定
-
-脚本不使用鼠标中键归正。每个 Point 到位后会：
-
-1. 使用路线末尾的 `orientation` 确定水平视角；
-2. 分多次将垂直镜头推向上限；
-3. 从统一上限向下移动固定量。
-
-相对鼠标移动有可能被游戏分帧截断。实测稳定默认值为向上推 `4` 次，每次之间等待 `100ms`，最后再等待 `150ms`，让第二、第三段连续路线也能重新获得一致的垂直基准。
-
-如果实机仍未到达上限，只需要在 JS 自定义配置中增加 **“镜头向上归一化重复次数”**；建议按 `1` 次逐步增加。通常不要先改下面的 `7500`，因为它控制的是到达上限以后向下放置的位置。
-
-设置项 **“从镜头垂直上限向下移动量（实测推荐 7500）”** 的默认值为 `7500`。当前新参数已在 `1920×1080` 和 `2560×1440` 下完成多轮实测；`3840×2160` 尚未验证这组新的镜头默认值。
-
-- 数值增大：镜头进一步向下；
-- 数值减小：镜头进一步向上；
-- 个别 Point 可在“特殊 Point 垂直下移覆盖”中填写，例如 `02=6100,05=5900`。
-
-如果水平位置需要单独微调，可使用“特殊 Point 水平微调覆盖”，格式同样为 `Point编号=数值`。
-
-## 分辨率、SIFT 与性能
-
-当前路线使用 `SIFT` 地图匹配。实机测试结果：
-
-- `1920×1080`：可以运行，但城内个别路线的方向修正更容易短暂摇摆；
-- `2560×1440`：当前推荐的准确性与性能甜点；
-- `3840×2160`：在运行流畅时，SIFT 判断最果断，但对硬件性能要求更高。
-
-在电脑能够稳定运行的前提下，更高分辨率可能为 SIFT 提供更多清晰特征。若 4K 导致帧率不稳，请降低画质或改用 `2560×1440`，稳定性优先于分辨率。
-
-检测到 `3840×2160` 或更高的游戏窗口时，脚本默认在三次“按 T 后、左键前”各额外等待 `800ms`。
-
-其他分辨率如果也存在 UI 动画尚未完成就点击的问题，可以打开 **“非 4K 分辨率也强制启用下方 T 后额外等待”**。开启后，1080p、1440p 等分辨率也会使用相同的 `800ms` 补偿；具体时间可以在下一项中调整，填 `0` 相当于不增加等待。4K 仍会自动启用，不需要手动勾选。
+检测到 `3840×2160` 或更高分辨率时，脚本默认在按 `T` 后额外等待 `800ms`。其他分辨率如果也出现界面动画未完成便点击的问题，可以手动启用相同时序补偿。
 
 ## 常用设置
 
-| 设置项 | 默认值 | 建议 |
+| 设置 | 默认值 | 建议 |
 | --- | ---: | --- |
 | 路线运行模式 | 按 Point 编号 | 首次测试保持默认 |
 | 每完成几个 Point 恢复热能 | 4 | 推荐保持 4 |
-| 热能恢复等待 | 1500ms | 恢复粒子未完成时再增加 |
-| 镜头向上归一化重复次数 | 4次 | 连续路线仍未到顶时逐次增加 |
-| 每次向上推镜头的间隔 | 100ms | 避免相对移动被游戏漏掉 |
-| 推到垂直上限后等待 | 150ms | 等镜头状态稳定 |
-| 从垂直上限向下移动量 | 7500 | 控制最终镜头向下多少 |
-| 第一次点赞等待 | 800ms | 点赞未刷新时适当增加 |
-| 第二次点赞等待 | 800ms | 点赞未刷新时适当增加 |
-| 非 4K 强制时序补偿 | 关闭 | 低分辨率动画较慢时手动打开 |
-| T 后额外等待 | 800ms | 4K 自动叠加，其他分辨率可强制启用 |
+| 镜头向上归一化次数 | 4 | 未到顶时逐次增加 |
+| 从垂直上限向下移动量 | 7500 | 控制最终放置位置 |
+| 两次点赞等待 | 各 800ms | 点赞未刷新时增加 |
+| T 后额外等待 | 800ms | 4K 自动启用 |
 
-其他延时项已有实测默认值，通常无需修改。
+## 故障排查与限制
 
-## 故障排查
-
-### 传送时总点在锚点下方
-
-如果原本可用的多条路线同时出现传送点击偏移，先完整退出并重启 BetterGI。实测中该问题重启后恢复，通常不是单条路线坐标错误。
-
-### 地图追踪反复摇摆
-
-确认截图方式为 `BitBlt`、小地图锁定方向，并优先保证稳定帧率。若 1080p 下仍明显摇摆，可尝试 `2560×1440`。
-
-### 车站 NPC 不存在或路线无收益
-
-Point 06、07、09 需要完成三个聚所任务的共同第四幕“雪后曙光”，可用成就“睁开被逸夜合拢的眼睛”验证；参考[影月月教程](https://www.bilibili.com/video/BV172bX6XEPm/)。Point 08 还需要在完成“雪后曙光”后的现实时间第二天完成“新雪来客”；参考[影月月教程](https://www.bilibili.com/video/BV1nqbX6fEDS/)。如果尚未完成“新雪来客”，请关闭 Point 08；脚本会在 Point 07 后直接执行包含 Point 09 工位的完整备用路线，不会再运行 `point09_continue.json`。
-
-### 脚本中途停止
-
-请先人工确认是否留下冰造物或仍停留在造物界面，再重新运行。
-
-## 已知限制
-
-- 当前版本不视觉识别实际点赞数量、热能数值或造物是否成功删除；
-- 日志中的“点赞等待完成”和“最终删除动作完成”表示固定动作及时序已完成，不代表视觉校验成功；
-- 路线依赖固定游戏环境、已解锁锚点与稳定帧率；
-- 新增路径会被脚本自动发现，但若希望在设置界面拥有独立开关，仍需同步更新 `settings.json`。
-
-## 更新记录
-
-### 1.1.0
-
-- 加入 Point 09，并将车站路线重构为 Point 07 → Point 08 → Point 09 连续组；
-- 加入 Point 08 关闭时从 Point 07 直达 Point 09 的备用路径；
-- 新增“高收益精简（6赞及以上）”运行模式，按路线组总点赞筛选；
-- 修正 Point 03 的预计点赞数为 4，确认 Point 04 为 2 赞，全部路线预计总数更新为 36；
-- 镜头默认值调整为向上4次、间隔100ms、到顶后等待150ms、再向下7500；
-- 更新“雪后曙光”和现实时间第二天任务“新雪来客”的路线前置说明。
-
-### 1.0.1
-
-- 镜头向上归一化由固定两次改为默认八次，并增加每次移动间隔；
-- 向上归一化次数和间隔可直接在 JS 自定义配置中调整；
-- 4K 继续自动启用 T 后延迟补偿；
-- 新增非 4K 分辨率手动强制启用相同补偿的开关；
-- 明确支持多个连续 `_continue` 路线组成不可拆分路线组。
-- 预留默认关闭的 Point 09 独立路线开关，加入路径文件后可直接启用。
-
-### 1.0.0
-
-- 收录 Point 01～08，预计总计 30 赞；
-- 支持逐路线开关、按点赞总数排序和 Point 07 → Point 08 连续路线；
-- 默认每 4 个 Point 恢复一次热能；
-- 使用固定垂直基准与默认 6000 镜头俯仰；
-- 增加 4K 的 T 后 800ms 自动补偿；
-- 每个 Point 完成两次点赞后最终删除造物并退出造物界面。
+- 多条路线同时出现传送点击偏移时，先完整退出并重启 BetterGI；
+- 地图追踪反复摇摆时，确认 `BitBlt`、小地图锁定方向和稳定帧率，条件允许时可尝试 `2560×1440`；
+- 脚本中途停止或报错后，请人工确认现场是否残留冰造物，并按 `T` 退出造物界面；
+- 日志中的预计点赞来自路线标签，脚本不会视觉识别实际点赞、热能数值或删除结果。
